requirements.txt
openai
python-dotenv
langchain
langchain-community
langchain_openai
fastapi
python-dotenv
faiss-cpu
pandas
sentence_transformers